Moved the hand hit mechanic to its own app state; Made SoundManager
